+++ Curve-finance-suite/src/agents/crossAssetSwap.ts
@@ -7,7 +7,7 @@ import {
 } from "forta-agent";
 
 import Web3 from "web3";
-import abi from "./stable-swap-abi";
+import abi from "../utils/stable-swap-abi";
 
 // @ts-ignore
 import abiDecoder from "abi-decoder";
@@ -15,23 +15,23 @@ abiDecoder.addABI(abi);
 
 export const web3 = new Web3();
 
-export const REMOVE_LIQUIDITY_IMBALANCE_SIGNATURE =
-  "RemoveLiquidiityImbalance(address, uint256[3], uint256[3],uint256, uint256)";
+export const CROSSCHAINSWAPSIGNATURE =
+  "TokenUpdate(uint256,address, address, uint256)";
 
 const createFinding = (alertID: string, address: string): Finding => {
   return Finding.fromObject({
-    name: "RemoveLiquidityImbalance Me funciton called",
-    description: "RemoveLiquidityImbalance Me funciton called on pool",
+    name: "CrossChainSwap Me funciton called",
+    description: "CrossChainSwap Me funciton called on pool",
     alertId: alertID,
     severity: FindingSeverity.Low,
     type: FindingType.Suspicious,
     metadata: {
-      data: address,
+      address: address,
     },
   });
 };
 
-export default function provideRemoveLiquidityImbalanceAgent(
+export default function provideCrossAssetSwap(
   alertID: string,
   address: string
 ): HandleTransaction {
@@ -40,7 +40,7 @@ export default function provideRemoveLiquidityImbalanceAgent(
 
     if (txEvent.addresses[address] == false) return findings;
 
-    if (txEvent.filterEvent(REMOVE_LIQUIDITY_IMBALANCE_SIGNATURE).length > 0) {
+    if (txEvent.filterEvent(CROSSCHAINSWAPSIGNATURE).length > 0) {
       findings.push(createFinding(alertID, address));
     }
