+++ Curve-finance-suite/src/agents/killme.ts
@@ -7,7 +7,7 @@ import {
 } from "forta-agent";
 
 import Web3 from "web3";
-import abi from "./stable-swap-abi";
+import abi from "../utils/stable-swap-abi";
 
 // @ts-ignore
 import abiDecoder from "abi-decoder";
@@ -15,23 +15,26 @@ abiDecoder.addABI(abi);
 
 export const web3 = new Web3();
 
-export const CROSSCHAINSWAPSIGNATURE =
-  "TokenUpdate(uint256,address, address, uint256)";
+export const killme = {
+  name: "kill_me",
+  outputs: [],
+  inputs: [],
+  stateMutability: "nonpayable",
+  type: "function",
+  gas: 38058,
+};
 
-const createFinding = (alertID: string, address: string): Finding => {
+const createFinding = (alertId: string): Finding => {
   return Finding.fromObject({
-    name: "CrossChainSwap Me funciton called",
-    description: "CrossChainSwap Me funciton called on pool",
-    alertId: alertID,
+    name: "Kill Me funciton called",
+    description: "Kill Me funciton called on pool",
+    alertId: alertId,
     severity: FindingSeverity.Low,
     type: FindingType.Suspicious,
-    metadata: {
-      address: address,
-    },
   });
 };
 
-export default function provideCrossAssetSwap(
+export default function provideKillMeAgent(
   alertID: string,
   address: string
 ): HandleTransaction {
@@ -40,8 +43,11 @@ export default function provideCrossAssetSwap(
 
     if (txEvent.addresses[address] == false) return findings;
 
-    if (txEvent.filterEvent(CROSSCHAINSWAPSIGNATURE).length > 0) {
-      findings.push(createFinding(alertID, address));
+    const data = abiDecoder.decodeMethod(txEvent.transaction.data);
+    if (!data) return findings;
+
+    if (data.name === "kill_me") {
+      findings.push(createFinding(alertID));
     }
 
     return findings;
