+++ Curve-finance-suite/src/agents/removeImbalanceLiquidity.ts
@@ -7,7 +7,7 @@ import {
 } from "forta-agent";
 
 import Web3 from "web3";
-import abi from "./stable-swap-abi";
+import abi from "../utils/stable-swap-abi";
 
 // @ts-ignore
 import abiDecoder from "abi-decoder";
@@ -15,26 +15,23 @@ abiDecoder.addABI(abi);
 
 export const web3 = new Web3();
 
-export const unkill = {
-  name: "unkill_me",
-  outputs: [],
-  inputs: [],
-  stateMutability: "nonpayable",
-  type: "function",
-  gas: 22195,
-};
+export const REMOVE_LIQUIDITY_IMBALANCE_SIGNATURE =
+  "RemoveLiquidiityImbalance(address, uint256[3], uint256[3],uint256, uint256)";
 
-const createFinding = (alertID: string): Finding => {
+const createFinding = (alertID: string, address: string): Finding => {
   return Finding.fromObject({
-    name: "UnKill Me funciton called",
-    description: "UnKill Me funciton called on pool",
+    name: "RemoveLiquidityImbalance Me funciton called",
+    description: "RemoveLiquidityImbalance Me funciton called on pool",
     alertId: alertID,
     severity: FindingSeverity.Low,
     type: FindingType.Suspicious,
+    metadata: {
+      data: address,
+    },
   });
 };
 
-export default function provideUnkillAgent(
+export default function provideRemoveLiquidityImbalanceAgent(
   alertID: string,
   address: string
 ): HandleTransaction {
@@ -43,11 +40,8 @@ export default function provideUnkillAgent(
 
     if (txEvent.addresses[address] == false) return findings;
 
-    const data = abiDecoder.decodeMethod(txEvent.transaction.data);
-    if (!data) return findings;
-
-    if (data.name === "unkill_me") {
-      findings.push(createFinding(alertID));
+    if (txEvent.filterEvent(REMOVE_LIQUIDITY_IMBALANCE_SIGNATURE).length > 0) {
+      findings.push(createFinding(alertID, address));
     }
 
     return findings;
