+++ Curve-finance-suite/src/utils/stable-swap-abi.ts
